+++ lib/foreman_tasks_core/output_collector_action.rb
@@ -1,0 +1,8 @@
+module ForemanTasksCore
+  class OutputCollectorAction < ::ForemanTasksCore::Runner::Action
+    def init_run
+      output[:result] = []
+      suspend
+    end
+  end
+end
